requirements.txt
opencv-python~=4.10.0.84
numpy~=2.0.2
pyproj~=3.6.1
rasterio~=1.4.3
geopy~=2.4.1
pillow